fefactor: 엔티티 속성값들 통일

+++ surbear-core/src/main/java/com/surbear/report/entity/PointHistoryEntity.java
@@ -30,7 +30,7 @@ public class PointHistoryEntity extends BaseTimeEntity {
     String description;
 
     @Column
-    Long paidPoint;
+    Integer paidPoint;
 
     @Builder.Default
     @Column
